+++ src/Completion/Content/ArrayContent.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace ByCerfrance\LlmApiLib\Completion\Content;
 
 use ArrayIterator;
+use InvalidArgumentException;
 use IteratorAggregate;
 use Override;
 use Traversable;
@@ -13,9 +14,34 @@ readonly class ArrayContent implements ContentInterface, IteratorAggregate
 {
     private array $contents;
 
-    public function __construct(ContentInterface ...$content)
+    public function __construct(ContentInterface|iterable|string|null ...$content)
     {
-        $this->contents = $content;
+        $this->contents = iterator_to_array($this->prepare($content), false);
+    }
+
+    private function prepare(iterable $contents): iterable
+    {
+        foreach ($contents as $content) {
+            if (null === $content) {
+                continue;
+            }
+
+            if (is_iterable($content)) {
+                yield from $this->prepare($content);
+                continue;
+            }
+
+            if (is_scalar($content)) {
+                yield new TextContent((string)$content);
+                continue;
+            }
+
+            if (!$content instanceof ContentInterface) {
+                throw new InvalidArgumentException('ArrayContent accept only ContentInterface or string types');
+            }
+
+            yield $content;
+        }
     }
 
     #[Override]