+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unrelease]
+## [1.2.0]
 
 ### Added
 
 - Support for **response_format** parameter in completions
 - Parameter `$role` for `CompletionInterface::getLastMessage()` method, to filter last message for given role
+
+### Changed
+
+- `ArrayContent` now accepts `ContentInterface`, iterable, null or string values
 
 ## [1.1.0] - 2025-06-17